Change: sepatation components
