+++ src/Button.jsx
@@ -1,0 +1,14 @@
+/* @jsx React.createElement */
+
+import React, { useState } from 'react';
+import ReactDOM from 'react-dom';
+
+function Button({ children }) {
+  return (
+    <>
+      <button type="button">{children}</button>
+    </>
+  );
+}
+
+export default Button;
