+++ src/Counter.jsx
@@ -1,0 +1,12 @@
+import React, { useState } from 'react';
+import ReactDOM from 'react-dom';
+
+function Counter({ count, onClick }) {
+  return (
+    <button type="button" onClick={onClick}>
+      click me! ({count})
+    </button>
+  );
+}
+
+export default Counter;
